requirements.txt
asgiref==3.10.0
Django==4.2.26
sqlparse==0.5.3
typing_extensions==4.15.0

# Django REST Framework
djangorestframework==3.14.0
django-filter==23.5
drf-spectacular==0.27.0
